[issue644] scoring functions compute doubles instead of ints (some future scoring criteria may want to compute ratios of certain properties or similar)

+++ src/search/merge_and_shrink/merge_scoring_function.h
@@ -16,7 +16,7 @@ protected:
 public:
     MergeScoringFunction();
     virtual ~MergeScoringFunction() = default;
-    virtual std::vector<int> compute_scores(
+    virtual std::vector<double> compute_scores(
         FactoredTransitionSystem &fts,
         const std::vector<std::pair<int, int>> &merge_candidates) = 0;
     // Overriding methods must set initialized to true.
